Added Footers
Added footers, set via the FO string, which print centered at the bottom of
each page. Regardless of whether it is defined or not, a trap springs the .DF,
draw footer, macro, two lines from the bottom of each page, breaking it and
creating two lines on the beginning of the new page, appropriately, after
drawing a centered footer title, if there is one defined.